# Conflicts:
#	python/boomer/common/prediction.pyx

+++ python/boomer/boosting/cpp/example_wise_losses.cpp
@@ -9,7 +9,7 @@ AbstractExampleWiseLoss::~AbstractExampleWiseLoss() {
 }
 
 void AbstractExampleWiseLoss::calculateGradientsAndHessians(AbstractRandomAccessLabelMatrix* labelMatrix,
-                                                            intp exampleIndex, const float64* predictedScores,
+                                                            uint32 exampleIndex, const float64* predictedScores,
                                                             float64* gradients, float64* hessians) {
 
 }
@@ -19,12 +19,12 @@ ExampleWiseLogisticLossImpl::~ExampleWiseLogisticLossImpl() {
 }
 
 void ExampleWiseLogisticLossImpl::calculateGradientsAndHessians(AbstractRandomAccessLabelMatrix* labelMatrix,
-                                                                intp exampleIndex, const float64* predictedScores,
+                                                                uint32 exampleIndex, const float64* predictedScores,
                                                                 float64* gradients, float64* hessians) {
-    intp numLabels = labelMatrix->numLabels_;
+    uint32 numLabels = labelMatrix->numLabels_;
     float64 sumOfExponentials = 1;
 
-    for (intp c = 0; c < numLabels; c++) {
+    for (uint32 c = 0; c < numLabels; c++) {
         uint8 trueLabel = labelMatrix->getLabel(exampleIndex, c);
         float64 expectedScore = trueLabel ? 1 : -1;
         float64 predictedScore = predictedScores[c];
@@ -34,9 +34,9 @@ void ExampleWiseLogisticLossImpl::calculateGradientsAndHessians(AbstractRandomAc
     }
 
     float64 sumOfExponentialsPow = pow(sumOfExponentials, 2);
-    intp i = 0;
+    uint32 i = 0;
 
-    for (intp c = 0; c < numLabels; c++) {
+    for (uint32 c = 0; c < numLabels; c++) {
         uint8 trueLabel = labelMatrix->getLabel(exampleIndex, c);
         float64 expectedScore = trueLabel ? 1 : -1;
         float64 predictedScore = predictedScores[c];
@@ -44,7 +44,7 @@ void ExampleWiseLogisticLossImpl::calculateGradientsAndHessians(AbstractRandomAc
         float64 tmp = (-expectedScore * exponential) / sumOfExponentials;
         gradients[c] = tmp;
 
-        for (intp c2 = 0; c2 < c; c2++) {
+        for (uint32 c2 = 0; c2 < c; c2++) {
             trueLabel = labelMatrix->getLabel(exampleIndex, c2);
             float64 expectedScore2 = trueLabel ? 1 : -1;
             float64 predictedScore2 = predictedScores[c2];
