+++ python/boomer/common/cpp/input_data.cpp
@@ -1,7 +1,7 @@
 #include "input_data.h"
 
 
-AbstractLabelMatrix::AbstractLabelMatrix(intp numExamples, intp numLabels) {
+AbstractLabelMatrix::AbstractLabelMatrix(uint32 numExamples, uint32 numLabels) {
     numExamples_ = numExamples;
     numLabels_ = numLabels;
 }
@@ -10,16 +10,16 @@ AbstractLabelMatrix::~AbstractLabelMatrix() {
 
 }
 
-AbstractRandomAccessLabelMatrix::AbstractRandomAccessLabelMatrix(intp numExamples, intp numLabels)
+AbstractRandomAccessLabelMatrix::AbstractRandomAccessLabelMatrix(uint32 numExamples, uint32 numLabels)
     : AbstractLabelMatrix(numExamples, numLabels) {
 
 }
 
-uint8 AbstractRandomAccessLabelMatrix::getLabel(intp exampleIndex, intp labelIndex) {
+uint8 AbstractRandomAccessLabelMatrix::getLabel(uint32 exampleIndex, uint32 labelIndex) {
     return 0;
 }
 
-DenseLabelMatrixImpl::DenseLabelMatrixImpl(intp numExamples, intp numLabels, const uint8* y)
+DenseLabelMatrixImpl::DenseLabelMatrixImpl(uint32 numExamples, uint32 numLabels, const uint8* y)
     : AbstractRandomAccessLabelMatrix(numExamples, numLabels) {
     y_ = y;
 }
@@ -28,12 +28,13 @@ DenseLabelMatrixImpl::~DenseLabelMatrixImpl() {
 
 }
 
-uint8 DenseLabelMatrixImpl::getLabel(intp exampleIndex, intp labelIndex) {
-    intp i = (exampleIndex * numLabels_) + labelIndex;
+uint8 DenseLabelMatrixImpl::getLabel(uint32 exampleIndex, uint32 labelIndex) {
+    uint32 i = (exampleIndex * numLabels_) + labelIndex;
     return y_[i];
 }
 
-DokLabelMatrixImpl::DokLabelMatrixImpl(intp numExamples, intp numLabels, std::shared_ptr<BinaryDokMatrix> dokMatrixPtr)
+DokLabelMatrixImpl::DokLabelMatrixImpl(uint32 numExamples, uint32 numLabels,
+                                       std::shared_ptr<BinaryDokMatrix> dokMatrixPtr)
     : AbstractRandomAccessLabelMatrix(numExamples, numLabels) {
     dokMatrixPtr_ = dokMatrixPtr;
 }
@@ -42,6 +43,6 @@ DokLabelMatrixImpl::~DokLabelMatrixImpl() {
 
 }
 
-uint8 DokLabelMatrixImpl::getLabel(intp exampleIndex, intp labelIndex) {
-    return dokMatrixPtr_.get()->getValue((uint32) exampleIndex, (uint32) labelIndex);
+uint8 DokLabelMatrixImpl::getLabel(uint32 exampleIndex, uint32 labelIndex) {
+    return dokMatrixPtr_.get()->getValue(exampleIndex, labelIndex);
 }
